+++ packages/time-select/src/time-picker/repeat-click.ts
